+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## ????
+- Github Actions for tests
+
 ## 0.0.3 - 2023-04-10
 - Provide default HTTP options
 - Default to HTTP/1.1