+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Static method, HTTP::g() for GET
 - New test for HTTP::g()
 - Static method, HTTP::p() for POST
+- New test for HTTP::p()
 - New simple PHP server that the tests run against, all tests have been updated to use it
 
 ## 0.0.3 - 2023-04-10